fix(idor): comparaisons d'ID en (int) partout (projets, portail employé, policies)
Même piège que les factures/devis : sur MySQL o2switch, une clé primaire
route-bound ($x->id) ressort en int et une clé étrangère ($y->..._id) en string,
donc '=== ' strict renvoie false et bloque à tort des utilisateurs légitimes (403).

Caste les deux côtés en (int) dans : TaskController, ProjectMemberController,
PortalExpense/Leave/Evaluation/SharedCalendar, policies (Room, ImportSession,
ProjectMember, HrEvent, SupportTicket), Organization, trait BelongsToUser.
Fail-safe : ne fait que rétablir l'accès légitime, n'ouvre aucun accès indu.

+++ app/Traits/BelongsToUser.php
@@ -85,7 +85,7 @@ trait BelongsToUser
     {
         $userId = $user instanceof User ? $user->id : $user;
 
-        return $this->user_id === $userId;
+        return (int) $this->user_id === (int) $userId;
     }
 
     /**
@@ -93,6 +93,6 @@ trait BelongsToUser
      */
     public function belongsToAuthUser(): bool
     {
-        return auth()->check() && $this->user_id === auth()->id();
+        return auth()->check() && (int) $this->user_id === (int) auth()->id();
     }
 }
